requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
ibm-cloud-sdk-core==3.16.0
ibm-vpc==0.13.0
idna==3.4
mysql-connector-python==8.0.31
protobuf==3.20.1
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
urllib3==1.26.13
